refactoring

+++ app/src/main/java/webServer/studentsProfiles/StudentProfileRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * StudentRepository
  */
-public interface StudentProfileRepository extends JpaRepository<StudentProfile, Long>{
+public interface StudentProfileRepository extends JpaRepository<StudentProfile, Integer>{
 
 	
 }
